gpu-gles: fix wrong long usage
No idea if there is any point touching this code but the warnings were
annoying.

+++ plugins/gpu-gles/gpuDraw.c
@@ -111,7 +111,7 @@ BOOL           bCheckMask=FALSE;
 int            iUseMask=0;
 int            iSetMask=0;
 unsigned short sSetMask=0;
-unsigned long  lSetMask=0;
+unsigned int   lSetMask=0;
 
 // drawing/coord vars
 
@@ -921,14 +921,14 @@ void offsetST(void)
 
 ///////////////////////////////////////////////////////// 
 
-void offsetScreenUpload(long Position)
+void offsetScreenUpload(int Position)
 {
  if(bDisplayNotSet)
   SetOGLDisplaySettings(1);
 
  if(Position==-1)
   {
-   long lmdx,lmdy;
+   int lmdx,lmdy;
 
    lmdx=xrUploadArea.x0;
    lmdy=xrUploadArea.y0;
