+++ plugins/gpu-gles/gpuDraw.h
@@ -64,7 +64,7 @@ unsigned short offsetline(void);
 #endif
 void offsetST(void);
 void offsetBlk(void);
-void offsetScreenUpload(long Position);
+void offsetScreenUpload(int Position);
 void assignTexture3(void);
 void assignTexture4(void);
 void assignTextureSprite(void);
