+++ plugins/gpu-gles/gpuPrim.h
@@ -39,8 +39,8 @@ extern EGLSurface surface;
 extern EGLDisplay display;
 #endif
 
-void UploadScreen (long Position);
-void PrepareFullScreenUpload (long Position);
+void UploadScreen (int Position);
+void PrepareFullScreenUpload (int Position);
 BOOL CheckAgainstScreen(short imageX0,short imageY0,short imageX1,short imageY1);
 BOOL CheckAgainstFrontScreen(short imageX0,short imageY0,short imageX1,short imageY1);
 BOOL FastCheckAgainstScreen(short imageX0,short imageY0,short imageX1,short imageY1);
